Add feature to print commit messages in dry run (#1)
Use the git process to print commit messages of the git repository
to the std output in a dry run. The format should be like so:

```
<hash> <head-message>
<hash> <head-message>
...
```

In addition the command should support to options, `--from` and
`--to`, those is expected to be the hash range within which each
commit should fall in. In case neither `--from` neither `--to` are
given the cli should print all the commits.

Resolves #1