Add operation to calc the tag range up to the HEAD (refs #12)
Implements an operation to calculate the starting point from which we should
include commits to the new release. This strating point is expected to be a
valid semver tag. In case the new release is a stable release the starting
point should be last in order stable tag otherwise either a stable or prerelease
is ok. No tags means all commits up to the HEAD should be taken into account.

Refs #12