Add a date next to the version in the changelog file (fixes #15)
By injecting the current release date next to the version number in each header
in the changelog, we improve readability of the file.

This change fixes partially the issue #15, we decide that it is a bit overwhelming
to add another command line argument (--date) just to format the release date.

Fixes #15