Refactor logger to reduce complexity (refs #17)
The linter is reporting about high complexities in the code within
some of the ops. The reason why is that currently each operation is
reporting to the console by checking if the global verbose option
has been set. This design leads to unnecessary condtional statements.
The refactoring of the logger is aiming to open the path to settle
down those complexities.

Refs #17